# OSSM-2436 Add readiness probe and improve first start