--- a/manifests-servicemesh/2.4.0/servicemeshoperator.v2.4.0.clusterserviceversion.yaml
+++ b/manifests-servicemesh/2.4.0/servicemeshoperator.v2.4.0.clusterserviceversion.yaml
@@ -17,7 +17,7 @@ metadata:
       The OpenShift Service Mesh Operator enables you to install, configure, and manage an instance of Red Hat OpenShift Service Mesh. OpenShift Service Mesh is based on the open source Istio project.
 
     containerImage: ${OSSM_OPERATOR_IMAGE}
-    createdAt: 2022-12-19T14:03:49EST
+    createdAt: 2023-04-06T13:11:56CEST
     support: Red Hat, Inc. 
     olm.skipRange: ">=1.0.2 <2.4.0-0"
     operators.openshift.io/infrastructure-features: '["Disconnected"]'
@@ -508,6 +508,8 @@ spec:
         spec:
 
           replicas: 1
+          strategy:
+            type: Recreate
           selector:
             matchLabels:
               name: istio-operator
@@ -569,6 +571,8 @@ spec:
                   ports:
                     - containerPort: 11999
                       name: validation
+                    - containerPort: 11200
+                      name: probes
                     - containerPort: 60000
                       name: metrics
                   command:
@@ -587,6 +591,11 @@ spec:
                       value: istio-operator
                       #        - name: ISTIO_CNI_IMAGE_PULL_SECRET
                       #          value: name-of-secret
+                  readinessProbe:
+                    httpGet:
+                      scheme: HTTP
+                      path: /readyz
+                      port: 11200
                   volumeMounts:
                     - name: operator-olm-config
                       mountPath: /etc/operator/olm