Before this fix, the first operator run looked like this:
```
$ kubectl get pods -w
NAME             READY   STATUS              RESTARTS   AGE
istio-operator   0/1     Pending             0          0s
istio-operator   0/1     Pending             0          0s
istio-operator   0/1     ContainerCreating   0          0s
istio-operator   0/1     ContainerCreating   0          2s
istio-operator   1/1     Running             0          17s    <-- shown as ready immediately after starting
istio-operator   0/1     Error               0          54s    <-- Error when restarting to pick up secrets
istio-operator   1/1     Running             1          61s    <-- immediately ready
```

After this fix, the operator starts like this:
```
NAME             READY   STATUS              RESTARTS   AGE
istio-operator   0/1     Pending             0          0s
istio-operator   0/1     Pending             0          0s
istio-operator   0/1     ContainerCreating   0          0s
istio-operator   0/1     ContainerCreating   0          3s
istio-operator   0/1     Running             0          11s    <-- running, but not yet ready
istio-operator   0/1     Completed           0          16s    <-- much faster restart and with no Error
istio-operator   0/1     Running             1          19s    <-- running, but not yet ready
istio-operator   1/1     Running             1          31s    <-- ready, because webhooks are ready
```