@luksa 2023-04-06
Has the API server's cert expired?